% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSED.R
\name{calcBIR}
\alias{calcBIR}
\title{Helper function to calculate BIR homologies}
\usage{
calcBIR(DT, bsg, window = 100, gapOpening = Inf, gapExtension = 4, ...)
}
\arguments{
\item{DT}{A data.table object containing fusion predictions.}

\item{bsg}{A BSgenome string or BSgenome object.}

\item{window}{A numeric value representing how upstream and downstream of
a fusion breakpoint within which to perform sequence alignment.}

\item{gapOpening}{A numeric value representing the cost for opening a gap
in the alignment.}

\item{gapExtension}{A numeric value representing the incremental cost for
extending the gap in alignment.}
}
\value{
A data.table object with putative homologies.
}
\description{
Returns a data.table for the fusion with the best scoring alignment
(or no alignment) for the BIR pathway.
}
