% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSED.R
\name{calcHomology}
\alias{calcHomology}
\title{Checks a fusion for hallmarks of microhomology error prone repairs.
Returns a data.table for the fusion with the best scoring alignment
(or no alignment) for each repair pathway.}
\usage{
calcHomology(
  DT,
  bsg,
  mechanism = c("ALL", "NHEJ", "SSA", "MMEJ", "BIR", "MMBIR"),
  window = NA,
  gapOpening = NA,
  gapExtension = NA,
  source = c("STAR", "ARRIBA", "ORIEN"),
  ...
)
}
\arguments{
\item{DT}{A data.table object representing a single fusion.}

\item{bsg}{A BSgenome string or BSgenome object.}

\item{mechanism}{A character vector representing which repair pathway to
investigate in the given fusion data.table.}

\item{window}{A numeric value representing how upstream and downstream of
a fusion breakpoint within which to perform sequence alignment.}

\item{gapOpening}{A numeric value representing the cost for opening a gap
in the alignment.}

\item{gapExtension}{A numeric value representing the incremental cost for
extending the gap in alignment.}

\item{source}{A character vector representing the fusion prediction source.}
}
\value{
A data.table object with fusion alignments for specified mechanism.
}
\description{
Checks a fusion for hallmarks of microhomology error prone repairs.
Returns a data.table for the fusion with the best scoring alignment
(or no alignment) for each repair pathway.
}
