% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSED.R
\name{calcPairwiseAlignment}
\alias{calcPairwiseAlignment}
\title{Helper function to calculate the optimal pairwise alignment of a fusion
of type, 'mechanism'.}
\usage{
calcPairwiseAlignment(
  DT,
  bsg,
  mechanism = c("NHEJ", "SSA", "MMEJ", "BIR", "MMBIR"),
  window = 4,
  gapOpening = Inf,
  gapExtension = Inf,
  ...
)
}
\arguments{
\item{DT}{A data.table object containing fusion predictions.}

\item{bsg}{A BSgenome string or BSgenome object.}

\item{mechanism}{A character vector representing which repair pathway to
investigate in the given fusion data.table.}

\item{window}{A numeric value representing how upstream and downstream of
a fusion breakpoint within which to perform sequence alignment.}

\item{gapOpening}{A numeric value representing the cost for opening a gap
in the alignment.}

\item{gapExtension}{A numeric value representing the incremental cost for
extending the gap in alignment.}
}
\value{
A data.table object with identified homologies.
}
\description{
Returns a data.table for the fusion with the best scoring alignment
(or no alignment) for each repair pathway.
}
