% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSED.R
\name{constructSTAR}
\alias{constructSTAR}
\title{Modify a valid STAR fusion data.table for downstream analysis}
\usage{
constructSTAR(x, window = 99)
}
\arguments{
\item{x}{A valid data.table object containing STAR predicted fusions.}

\item{window}{A numeric value by which to expand the region around the
breakpoint}
}
\value{
A fusion data.table object
}
\description{
Modify a valid STAR fusion data.table for downstream analysis
}
