% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSED.R
\name{.validARRIBA}
\alias{.validARRIBA}
\title{Checks if provided object is an ARRIBA derived fusion prediction data.table}
\usage{
.validARRIBA(x)
}
\arguments{
\item{x}{A data.table object containing ARRIBA predicted fusions.}
}
\value{
A warning if the object does not appear to be from ARRIBA
}
\description{
Checks if provided object is an ARRIBA derived fusion prediction data.table
}
