% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSED.R
\name{.validBSgenome}
\alias{.validBSgenome}
\title{Checks if provided genome is an available BSgenome object,
and if so, returns it. If the library is not installed, it issues a warning
and returns NULL.}
\usage{
.validBSgenome(bsg)
}
\arguments{
\item{bsg}{A BSgenome compatible genome string or BSgenome object.}
}
\value{
A BSgenome object if installed.
}
\description{
Checks if provided genome is an available BSgenome object,
and if so, returns it. If the library is not installed, it issues a warning
and returns NULL.
}
