% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSED.R
\name{.validORIEN}
\alias{.validORIEN}
\title{Checks if provided object is an ORIEN derived fusion prediction data.table}
\usage{
.validORIEN(x)
}
\arguments{
\item{x}{A data.table object containing ORIEN predicted fusions.}
}
\value{
A warning if the object does not appear to be from ORIEN
}
\description{
Checks if provided object is an ORIEN derived fusion prediction data.table
}
