% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSED.R
\name{.validSTAR}
\alias{.validSTAR}
\title{Checks if provided object is a STAR derived fusion prediction data.table}
\usage{
.validSTAR(x)
}
\arguments{
\item{x}{A data.table object containing STAR predicted fusions.}
}
\value{
A warning if the object does not appear to be from STAR.
}
\description{
Checks if provided object is a STAR derived fusion prediction data.table
}
