% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSED.R
\docType{package}
\name{fused}
\alias{fused}
\title{FUSED R package}
\description{
FUSED is a collection of functions to determine the presence of hallmarks
of common DNA repair mechanisms including NHEJ, MMEJ, SSA, BIR, and MMBIR.
}
\references{
\url{http://github.com/databio/fused}
}
\author{
Jason Smith, Ben Morris
}
