% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSED.R
\name{plotHomology}
\alias{plotHomology}
\title{Plot the fusions that pass by mechanism.}
\usage{
plotHomology(DT)
}
\arguments{
\item{DT}{A data.table object representing a single fusion.}
}
\value{
A ggplot object.
}
\description{
Plot the fusions that pass by mechanism.
}
