% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSED.R
\name{splitDataTable}
\alias{splitDataTable}
\title{Splits a data.table object into a list of data.tables based on an
individual column. Returns a list of data.tables if column to split
is present, otherwise returns NULL.}
\usage{
splitDataTable(DT, split_factor)
}
\arguments{
\item{DT}{A data.table object.}

\item{split_factor}{A column name in DT to use as a factor to split and
name list elements. Can be a character vector
or an integer.}
}
\value{
A list of data.tables.
}
\description{
Splits a data.table object into a list of data.tables based on an
individual column. Returns a list of data.tables if column to split
is present, otherwise returns NULL.
}
