% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUSED.R
\name{theme_FUSED}
\alias{theme_FUSED}
\title{A standardized ggplot theme for FUSED plots}
\usage{
theme_FUSED(base_family = "sans", ...)
}
\description{
A standardized ggplot theme for FUSED plots
}
\examples{
theme_FUSED()
}
\keyword{ggplot2}
\keyword{theme}
